requirements.txt
Flask==3.0.3
gunicorn==22.0.0
supabase==2.5.3
gotrue==2.5.2
python-dotenv==1.0.1
Werkzeug==3.0.3
pytz==2024.1
